Configured the Solution to be able to use EFCore
