+++ ConsoleApp/ConsoleApp.csproj
@@ -9,7 +9,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ContabilidadDbContext\DbContextLibrary.csproj" />
-    <ProjectReference Include="..\Shared\Shared.csproj" />
   </ItemGroup>
 
 </Project>
