+++ ManejoContable/appsettings.json
@@ -1,0 +1,7 @@
+{
+  
+
+  "ConnectionStrings": {
+    "Contabilidad": "Data Source=localhost; Initial Catalog=ContabilidadDb; Trusted_Connection=True"
+  }
+}
