--- Shared/appsettings.json
@@ -1,14 +1,0 @@
-{
-  "exclude": [
-    "**/bin",
-    "**/bower_components",
-    "**/jspm_packages",
-    "**/node_modules",
-    "**/obj",
-    "**/platforms"
-  ],
-
-  "ConnectionStrings": {
-    "Contabilidad": "Data Source=localhost; Initial Catalog=ContabilidadDb; Trusted_Connection=True"
-  }
-}
